# specs/001-medaware-monorepo/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MedAware API
  description: Reddit-style medical Q&A MVP. No auth; x-user-id header identifies acting user for writes.
  version: 0.1.0

servers:
  - url: http://localhost:3001
    description: Local API

tags:
  - name: Users
    description: User management (read-only for MVP)
  - name: Questions
    description: Medical questions
  - name: Answers
    description: Doctor answers

paths:
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'

  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleUserResponse'
        '404':
          description: User not found

  /questions:
    get:
      tags: [Questions]
      summary: List questions with filters and pagination
      operationId: listQuestions
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListQuestionsResponse'
    post:
      tags: [Questions]
      summary: Create question (requires x-user-id, must be PATIENT)
      operationId: createQuestion
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionInput'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleQuestionResponse'
        '400':
          description: Validation error
        '401':
          description: Missing or invalid x-user-id
        '403':
          description: User is not a patient

  /questions/{id}:
    get:
      tags: [Questions]
      summary: Get question by ID (includes answers)
      operationId: getQuestion
      parameters:
        - $ref: '#/components/parameters/QuestionIdPath'
      responses:
        '200':
          description: Question with answers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionWithAnswersResponse'
        '404':
          description: Question not found
    patch:
      tags: [Questions]
      summary: Update question (author only)
      operationId: updateQuestion
      parameters:
        - $ref: '#/components/parameters/QuestionIdPath'
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionInput'
      responses:
        '200':
          description: Question updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleQuestionResponse'
        '403':
          description: Not the author
        '404':
          description: Question not found
    delete:
      tags: [Questions]
      summary: Delete question (author only)
      operationId: deleteQuestion
      parameters:
        - $ref: '#/components/parameters/QuestionIdPath'
        - $ref: '#/components/parameters/XUserIdHeader'
      responses:
        '204':
          description: Question deleted
        '403':
          description: Not the author
        '404':
          description: Question not found

  /questions/{id}/answers:
    post:
      tags: [Answers]
      summary: Create answer (requires x-user-id, must be verified doctor)
      operationId: createAnswer
      parameters:
        - $ref: '#/components/parameters/QuestionIdPath'
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnswerInput'
      responses:
        '201':
          description: Answer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleAnswerResponse'
        '400':
          description: Validation error or question closed
        '401':
          description: Missing or invalid x-user-id
        '403':
          description: User is not a verified doctor
        '404':
          description: Question not found

  /answers/{id}:
    patch:
      tags: [Answers]
      summary: Update answer (owner doctor only)
      operationId: updateAnswer
      parameters:
        - $ref: '#/components/parameters/AnswerIdPath'
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnswerInput'
      responses:
        '200':
          description: Answer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleAnswerResponse'
        '403':
          description: Not the answer owner
        '404':
          description: Answer not found
    delete:
      tags: [Answers]
      summary: Delete answer (owner doctor only)
      operationId: deleteAnswer
      parameters:
        - $ref: '#/components/parameters/AnswerIdPath'
        - $ref: '#/components/parameters/XUserIdHeader'
      responses:
        '204':
          description: Answer deleted
        '403':
          description: Not the answer owner
        '404':
          description: Answer not found

components:
  parameters:
    XUserIdHeader:
      name: x-user-id
      in: header
      description: UUID of acting user (required for create/update/delete)
      required: false
      schema: { type: string, format: uuid }
    UserIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    QuestionIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    AnswerIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }

  schemas:
    ApiResponse:
      type: object
      properties:
        data: {}
        meta:
          type: object
          description: Optional pagination/metadata

    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        role: { type: string, enum: [PATIENT, DOCTOR] }
        is_verified_doctor: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    ListUsersResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/User' }

    SingleUserResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/User' }

    Question:
      type: object
      properties:
        id: { type: string, format: uuid }
        author_id: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        tags: { type: array, items: { type: string } }
        status: { type: string, enum: [OPEN, CLOSED] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    CreateQuestionInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 5, maxLength: 120 }
        body: { type: string, minLength: 20, maxLength: 5000 }
        tags: { type: array, items: { type: string }, maxItems: 5 }

    UpdateQuestionInput:
      type: object
      properties:
        title: { type: string, minLength: 5, maxLength: 120 }
        body: { type: string, minLength: 20, maxLength: 5000 }
        tags: { type: array, items: { type: string }, maxItems: 5 }
        status: { type: string, enum: [OPEN, CLOSED] }

    Answer:
      type: object
      properties:
        id: { type: string, format: uuid }
        question_id: { type: string, format: uuid }
        doctor_id: { type: string, format: uuid }
        body: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    CreateAnswerInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 20, maxLength: 5000 }

    UpdateAnswerInput:
      type: object
      properties:
        body: { type: string, minLength: 20, maxLength: 5000 }

    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }

    ListQuestionsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Question' }
            meta:
              $ref: '#/components/schemas/PaginationMeta'

    SingleQuestionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Question' }

    QuestionWithAnswersResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/Question'
                - type: object
                  properties:
                    answers:
                      type: array
                      items: { $ref: '#/components/schemas/Answer' }

    SingleAnswerResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Answer' }